+++ pcs_test/tier0/lib/cib/resource/test_primitive_validate.py
@@ -592,6 +592,22 @@ class ValidateResourceInstanceAttributesCreateSelfValidation(TestCase):
         self.agent_self_validation_mock.return_value = True, []
         self.cmd_runner = mock.Mock()
 
+    def test_disabled(self):
+        attributes = {"required": "value"}
+        facade = _fixture_ocf_agent()
+        self.assertEqual(
+            primitive.validate_resource_instance_attributes_create(
+                self.cmd_runner,
+                facade,
+                attributes,
+                etree.Element("resources"),
+                force=False,
+                enable_agent_self_validation=False,
+            ),
+            [],
+        )
+        self.agent_self_validation_mock.assert_not_called()
+
     def test_success(self):
         attributes = {"required": "value"}
         facade = _fixture_ocf_agent()
@@ -602,6 +618,7 @@ class ValidateResourceInstanceAttributesCreateSelfValidation(TestCase):
                 attributes,
                 etree.Element("resources"),
                 force=False,
+                enable_agent_self_validation=True,
             ),
             [],
         )
@@ -621,6 +638,7 @@ class ValidateResourceInstanceAttributesCreateSelfValidation(TestCase):
                 attributes,
                 etree.Element("resources"),
                 force=True,
+                enable_agent_self_validation=True,
             ),
             [],
         )
@@ -642,6 +660,7 @@ class ValidateResourceInstanceAttributesCreateSelfValidation(TestCase):
                 attributes,
                 etree.Element("resources"),
                 force=False,
+                enable_agent_self_validation=True,
             ),
             [
                 fixture.error(
@@ -667,6 +686,7 @@ class ValidateResourceInstanceAttributesCreateSelfValidation(TestCase):
                 attributes,
                 etree.Element("resources"),
                 force=False,
+                enable_agent_self_validation=True,
             ),
             [],
         )
@@ -686,6 +706,7 @@ class ValidateResourceInstanceAttributesCreateSelfValidation(TestCase):
                 attributes,
                 etree.Element("resources"),
                 force=False,
+                enable_agent_self_validation=True,
             ),
             [],
         )
@@ -701,6 +722,7 @@ class ValidateResourceInstanceAttributesCreateSelfValidation(TestCase):
                 attributes,
                 etree.Element("resources"),
                 force=False,
+                enable_agent_self_validation=True,
             ),
             [],
         )
@@ -716,6 +738,7 @@ class ValidateResourceInstanceAttributesCreateSelfValidation(TestCase):
                 attributes,
                 etree.Element("resources"),
                 force=False,
+                enable_agent_self_validation=True,
             ),
             [
                 fixture.error(
@@ -1275,6 +1298,24 @@ class ValidateResourceInstanceAttributesUpdateSelfValidation(TestCase):
             etree.SubElement(nvset_el, "nvpair", dict(name=name, value=value))
         return resources_el
 
+    def test_disabled(self):
+        old_attributes = {"required": "old_value"}
+        new_attributes = {"required": "new_value"}
+        facade = _fixture_ocf_agent()
+        self.assertEqual(
+            primitive.validate_resource_instance_attributes_update(
+                self.cmd_runner,
+                facade,
+                new_attributes,
+                self._NAME,
+                self._fixture_resources(old_attributes),
+                force=False,
+                enable_agent_self_validation=False,
+            ),
+            [],
+        )
+        self.agent_self_validation_mock.assert_not_called()
+
     def test_success(self):
         old_attributes = {"required": "old_value"}
         new_attributes = {"required": "new_value"}
@@ -1287,6 +1328,7 @@ class ValidateResourceInstanceAttributesUpdateSelfValidation(TestCase):
                 self._NAME,
                 self._fixture_resources(old_attributes),
                 force=False,
+                enable_agent_self_validation=True,
             ),
             [],
         )
@@ -1318,6 +1360,7 @@ class ValidateResourceInstanceAttributesUpdateSelfValidation(TestCase):
                 self._NAME,
                 self._fixture_resources(old_attributes),
                 force=True,
+                enable_agent_self_validation=True,
             ),
             [],
         )
@@ -1354,6 +1397,7 @@ class ValidateResourceInstanceAttributesUpdateSelfValidation(TestCase):
                 self._NAME,
                 self._fixture_resources(old_attributes),
                 force=False,
+                enable_agent_self_validation=True,
             ),
             [
                 fixture.error(
@@ -1391,6 +1435,7 @@ class ValidateResourceInstanceAttributesUpdateSelfValidation(TestCase):
                 self._NAME,
                 self._fixture_resources(old_attributes),
                 force=False,
+                enable_agent_self_validation=True,
             ),
             [],
         )
@@ -1422,6 +1467,7 @@ class ValidateResourceInstanceAttributesUpdateSelfValidation(TestCase):
                 self._NAME,
                 self._fixture_resources(old_attributes),
                 force=False,
+                enable_agent_self_validation=True,
             ),
             [],
         )
@@ -1439,6 +1485,7 @@ class ValidateResourceInstanceAttributesUpdateSelfValidation(TestCase):
                 self._NAME,
                 self._fixture_resources(old_attributes),
                 force=False,
+                enable_agent_self_validation=True,
             ),
             [],
         )
@@ -1456,6 +1503,7 @@ class ValidateResourceInstanceAttributesUpdateSelfValidation(TestCase):
                 self._NAME,
                 self._fixture_resources(old_attributes),
                 force=False,
+                enable_agent_self_validation=True,
             ),
             [
                 fixture.error(
@@ -1482,6 +1530,7 @@ class ValidateResourceInstanceAttributesUpdateSelfValidation(TestCase):
                 self._NAME,
                 self._fixture_resources(old_attributes),
                 force=False,
+                enable_agent_self_validation=True,
             ),
             [
                 fixture.warn(